+++ Nos/Views/AppView.swift
@@ -12,21 +12,23 @@ struct AppView: View {
     @EnvironmentObject private var appController: AppController
     
     @State var isCreatingNewPost = false
+    
+    @State var showNewPost = false
 
     @EnvironmentObject var router: Router
     
     @Environment(\.managedObjectContext) private var viewContext
     
-    @State
-    private var showingOptions = false
+    @State private var showingOptions = false
+    @State private var lastSelectedTab = Destination.home
     
     /// An enumeration of the destinations for AppView.
-    enum Destination: String, Hashable {
+    enum Destination: String, Hashable, Equatable {
         case home
         case discover
-        case relays
         case notifications
         case newNote
+        case profile
         
         var label: some View {
             switch self {
@@ -34,12 +36,12 @@ struct AppView: View {
                 return Text(Localized.homeFeed.string)
             case .discover:
                 return Localized.discover.view
-            case .relays:
-                return Text(Localized.relays.string)
             case .notifications:
                 return Localized.notifications.view
             case .newNote:
                 return Localized.newNote.view
+            case .profile:
+                return Localized.profile.view
             }
         }
         
@@ -49,12 +51,12 @@ struct AppView: View {
                 return Localized.homeFeed.string
             case .discover:
                 return Localized.discover.string
-            case .relays:
-                return Localized.relays.string
             case .notifications:
                 return Localized.notifications.string
             case .newNote:
                 return Localized.newNote.string
+            case .profile:
+                return Localized.profile.string
             }
         }
     }
@@ -67,34 +69,89 @@ struct AppView: View {
             } else {
                 TabView(selection: $router.selectedTab) {
                     HomeFeedView(user: CurrentUser.author(in: viewContext))
-                        .tabItem { Label(Localized.homeFeed.string, systemImage: "house") }
+                        .tabItem {
+                            VStack {
+                                let text = Localized.homeFeed.view
+                                if $router.selectedTab.wrappedValue == .home {
+                                    Image.tabIconHomeSelected
+                                    text
+                                } else {
+                                    Image.tabIconHome
+                                    text.foregroundColor(.secondaryTxt)
+                                }
+                            }
+                        }
                         .toolbarBackground(Color.cardBgBottom, for: .tabBar)
                         .tag(Destination.home)
                     
                     DiscoverView()
-                        .tabItem { Label(Localized.discover.string, systemImage: "magnifyingglass") }
+                        .tabItem {
+                            VStack {
+                                let text = Localized.discover.view
+                                if $router.selectedTab.wrappedValue == .discover {
+                                    Image.tabIconEveryoneSelected
+                                    text.foregroundColor(.textColor)
+                                } else {
+                                    Image.tabIconEveryone
+                                    text.foregroundColor(.secondaryTxt)
+                                }
+                            }
+                        }
                         .toolbarBackground(Color.cardBgBottom, for: .tabBar)
                         .tag(Destination.discover)
                     
-                    NewNoteView(isPresented: .constant(true))
+                    VStack {}
                         .tabItem {
-                            Label(Localized.newNote.string, systemImage: "plus.circle")
+                            VStack {
+                                Image.newPostButton
+                                Localized.post.view
+                            }
                         }
-                        .toolbarBackground(Color.cardBgBottom, for: .tabBar)
-                        .tag(Destination.newNote)
+                    .tag(Destination.newNote)
                     
                     NotificationsView(user: CurrentUser.author(in: viewContext))
-                        .tabItem { Label(Localized.notifications.string, systemImage: "bell") }
+                        .tabItem {
+                            VStack {
+                                let text = Localized.notifications.view
+                                if $router.selectedTab.wrappedValue == .notifications {
+                                    Image.tabIconNotificationsSelected
+                                    text.foregroundColor(.textColor)
+                                } else {
+                                    Image.tabIconNotifications
+                                    text.foregroundColor(.secondaryTxt)
+                                }
+                            }
+                        }
                         .toolbarBackground(Color.cardBgBottom, for: .tabBar)
                         .tag(Destination.notifications)
                     
-                    RelayView(author: CurrentUser.author(in: viewContext)!)
-                        .tabItem {
-                            Label(Localized.relays.string, systemImage: "antenna.radiowaves.left.and.right")
-                        }
-                        .toolbarBackground(Color.cardBgBottom, for: .tabBar)
-                        .tag(Destination.relays)
+                    if let author = CurrentUser.author(in: viewContext) {
+                        ProfileTab(author: author, path: $router.profilePath)
+                            .tabItem {
+                                VStack {
+                                    let text = Localized.profile.view
+                                    if $router.selectedTab.wrappedValue == .profile {
+                                        text.foregroundColor(.textColor)
+                                    } else {
+                                        text.foregroundColor(.secondaryTxt)
+                                    }
+                                }
+                            }
+                            .toolbarBackground(Color.cardBgBottom, for: .tabBar)
+                            .tag(Destination.profile)
+                    }
                 }
+                .onChange(of: router.selectedTab) { newTab in
+                    if newTab == Destination.newNote {
+                        showNewPost = true
+                        router.selectedTab = lastSelectedTab
+                    } else if !showNewPost {
+                        lastSelectedTab = newTab
+                    }
+                }
+                .sheet(isPresented: $showNewPost, content: {
+                    NewNoteView(isPresented: $showNewPost)
+                })
                 
                 SideMenu(
                     menuWidth: UIScreen.main.bounds.width / 1.3,
@@ -105,14 +162,18 @@ struct AppView: View {
             }
         }
         .onAppear(perform: appController.configureCurrentState)
-        .onAppear {
+        .task {
             let nosAppearance = UINavigationBarAppearance()
             nosAppearance.titleTextAttributes = [.foregroundColor: UIColor.primaryTxt]
             nosAppearance.largeTitleTextAttributes = [.foregroundColor: UIColor.primaryTxt]
             UINavigationBar.appearance().standardAppearance = nosAppearance
             UINavigationBar.appearance().compactAppearance = nosAppearance
             UINavigationBar.appearance().scrollEdgeAppearance = nosAppearance
+            
+            UITabBar.appearance().unselectedItemTintColor = .secondaryText
+            UITabBar.appearance().tintColor = .primaryTxt
         }
+        .accentColor(.primaryTxt)
     }
 }