+++ Nos/Views/FollowCard.swift
@@ -31,7 +31,7 @@ struct FollowCard: View {
                         Text(author.safeName)
                             .lineLimit(1)
                             .font(.subheadline)
-                            .foregroundColor(Color.secondaryTxt)
+                            .foregroundColor(Color.primaryTxt)
                             .multilineTextAlignment(.leading)
                             .frame(maxWidth: .infinity, alignment: .leading)
                         if author.muted {
@@ -42,6 +42,7 @@ struct FollowCard: View {
                         Spacer()
                         if let currentUser = CurrentUser.author {
                             FollowButton(currentUserAuthor: currentUser, author: author)
+                                .padding(10)
                         }
                     }
                 }
@@ -59,16 +60,6 @@ struct FollowCard: View {
         )
         .listRowInsets(EdgeInsets())
         .cornerRadius(cornerRadius)
-        .padding(padding)
-    }
-
-    var padding: EdgeInsets {
-        switch style {
-        case .golden:
-            return EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0)
-        case .compact:
-            return EdgeInsets(top: 15, leading: 0, bottom: 0, trailing: 0)
-        }
     }
 
     var cornerRadius: CGFloat {
@@ -76,7 +67,7 @@ struct FollowCard: View {
         case .golden:
             return 15
         case .compact:
-            return 20
+            return 15
         }
     }
 }